Working version that filters along with defined types
Removed CSS styling for some items in favor of using the PrimeNG classes to define their height, width, margin, etc. Also defined types for API response and the items that would be used for the application. Items will filter and reset themselves successfully. Need to adjust so that the selection boxes do not move whenever the filter occurs.